scriptops-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.7.4
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
pyyaml==6.0.1
apscheduler==3.10.4
